feat(interest): ineligibility warning
